Delete LICENSE
The license is not necessary because this program has so little code that it does not need to be anything official.  Anyone who wants to use this code may use it, and anyone who wishes to modify it may modify it.  There is, therefore, no need for a license to be included.